+++ source/renderer/zettlr-dialog.js
@@ -199,7 +199,11 @@ class ZettlrDialog {
         // can be previewed before it is decided what to do with it.
         obj.img = clipboard.readImage().toDataURL()
         obj.size = clipboard.readImage().getSize()
+        obj.aspect = clipboard.readImage().getAspectRatio()
         if (clipboard.readText().length > 0) {
+          // If you copy an image from the web, the browser sometimes inserts
+          // the original URL to it as text into the clipboard. In this case
+          // we've already got a good image name!
           obj.imageName = path.basename(clipboard.readText(), path.extname(clipboard.readText()))
         } else {
           // In case there is no potential basename we could extract, simply
@@ -348,6 +352,27 @@ class ZettlrDialog {
       }
       this._chart = new Chart(document.getElementById('canvas').getContext('2d'), config)
     } // END initiate canvas
+
+    // Initiate the paste-image dialog
+    if (this._dlg === 'paste-image') {
+      // We need to trigger a replace manually, b/c pasting raw image data into
+      // the src-attribute of an img does not trigger the onLoad-event listener.
+      setTimeout(() => { this._place() }, 10)
+
+      // Activate buttons
+      $('#save-cwd, #save-other').on('click', (e) => {
+        // The content is either save-cwd or save-other
+        global.ipc.send('save-image-from-clipboard', {
+          'mode': $(e.target).attr('id'),
+          'name': $('#img-name').val(),
+          'width': parseInt($('#img-width').val() || 0), // Resize to width, 0 indicates no change
+          'height': parseInt($('#img-height').val() || 0) // Resize to height, 0 indicates no change
+        })
+
+        // Now close the dialog
+        this.close()
+      }) // End activate buttons
+    }
 
     return this
   }